pyproject.toml
[project]
name = "agentic-deepfake-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "dlib-bin>=19.24.0",
    "fastapi>=0.100.0",
    "gradio>=6.2.0",
    "keras>=3.13.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "orch>=0.0.2",
    "pillow>=10.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.28.0",
    "tensorflow>=2.20.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.65.0",
    "transformers>=4.57.3",
    "uvicorn>=0.23.0",
    "yolov9>=0.0.1",
]
